move logic to seperate wrapper
