+++ src/AppContainer.tsx
@@ -3,16 +3,19 @@ import { StatusBar } from 'react-native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 
 import { RootNavigator } from './RootNavigator';
+import { SaveActiveAppState } from './SaveActiveAppState';
 import { AppGate } from './components';
 
 const AppContainer = () => {
   return (
-    <AppGate>
-      <SafeAreaProvider>
-        <StatusBar backgroundColor="transparent" />
-        <RootNavigator />
-      </SafeAreaProvider>
-    </AppGate>
+    <SaveActiveAppState>
+      <AppGate>
+        <SafeAreaProvider>
+          <StatusBar backgroundColor="transparent" />
+          <RootNavigator />
+        </SafeAreaProvider>
+      </AppGate>
+    </SaveActiveAppState>
   );
 };
